+++ Intuit.TSheets/Model/Filters/JobcodeFilter.cs
@@ -40,7 +40,7 @@ namespace Intuit.TSheets.Model.Filters
         [JsonConverter(typeof(EnumerableToCsvConverter))]
         [JsonSchema(JsonObjectType.String)]
         [JsonProperty("ids")]
-        public IEnumerable<int> Ids { get; set; }
+        public IEnumerable<long> Ids { get; set; }
 
         /// <summary>
         /// Gets or sets the jobcode parent ids you'd like to filter on.
@@ -48,7 +48,7 @@ namespace Intuit.TSheets.Model.Filters
         [JsonConverter(typeof(EnumerableToCsvConverter))]
         [JsonSchema(JsonObjectType.String)]
         [JsonProperty("parent_ids")]
-        public IEnumerable<int> ParentIds { get; set; }
+        public IEnumerable<long> ParentIds { get; set; }
 
         /// <summary>
         /// Gets or sets the wildcard-able filter on the jobcode name.
